Update CSS builds w/ repo & submodule dependency changes
First, update inline documentation in ascc-custom.less by providing greater clarity and fixing a mistake in the file header comment. Next, update the submodule for the DAESA-wide CSS development project to incorporate changes in the dependencies it contains into the superproject. Finally, rebuild the website's custom CSS production files to integrate all of these dependency changes into the final stylesheet that is applied to the site via WSUWP.