Update dev dependencies for and builds of custom CSS code
First, add style rules to ascc-custom.less for a responsive two column layout for the list of ASCC resources for students during social distancing associated with the COVID-19 outbreak, and update rules for glass button styled hyperlinks. Then rebuild the files containing custom CSS code that is ready to be deployed on the website associated with this project.